+++ R/aaa.R
@@ -1,8 +1,3 @@
-# better try
-tri <- function(expr) try_fetch(expr, error = function(e) e)
-is.condition <- function(x) inherits(x, 'condition')
-is.error_cond <- function(x) is.condition(x) && inherits(x, 'error')
-
 check_function_args <- function(x, args, ..., allow_null = FALSE, arg = caller_arg(x), call = caller_env()) {
   check_function(x, ..., allow_null = allow_null, arg = arg, call = call)
   if (!is.null(x) && !all(args %in% fn_fmls_names(x))) {