+++ R/ressource_route.R
@@ -173,14 +173,7 @@ ressource_route <- function(..., default_file = 'index.html', default_ext = 'htm
         }
       }
       if (!is.null(finalize)) {
-        success <- tri(finalize(request, response, ...))
-        if (reqres::is_reqres_problem(success)) {
-          reqres::handle_problem(response, success)
-          return(FALSE)
-        } else if (is_condition(success)) {
-          response$status_with_text(500L, clear_headers = TRUE)
-          return(FALSE)
-        }
+        finalize(request, response, ...)
       }
       continue
     })